lab2 queries

+++ database/lab2/Lab2/Lab2/Models/Post.cs
@@ -19,6 +19,6 @@ namespace Lab2.Models
         public string Title { get; set; }
         [Range(0, 100)]
         [Required]
-        public double Pulse { get; set; }
+        public int Pulse { get; set; }
     }
 }
